pyproject.toml
[project]
name = "termikita"
version = "0.1.0"
description = "Native macOS terminal emulator with full Unicode/Vietnamese support"
requires-python = ">=3.12"

dependencies = [
    "pyobjc-framework-Cocoa>=10.0",
    "pyobjc-framework-CoreText>=10.0",
    "pyobjc-framework-Quartz>=10.0",
    "pyte>=0.8.0",
]

[project.optional-dependencies]
dev = ["pytest", "ruff", "mypy"]
build = ["py2app"]

[project.scripts]
termikita = "termikita:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E501"]
